// src/types/Color.ts
export type Color =
  | 'black'
  | 'gold'
  | 'yellow'
  | 'green'
  | 'midnightgreen'
  | 'silver'
  | 'spacegray'
  | 'red'
  | 'white'
  | 'purple'
  | 'coral'
  | 'rosegold'
  | 'midnight'
  | 'spaceblack'
  | 'blue'
  | 'pink'
  | 'sierrablue'
  | 'graphite'
  | 'space gray'
  | 'space-gray'
  | 'rose gold'
  | 'rose-gold'
  | 'sky blue'
  | 'sky-blue'
  | 'starlight';